% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowWAFOMSobol.R
\name{lowWAFOMSobol.dimF2MinMax}
\alias{lowWAFOMSobol.dimF2MinMax}
\title{get minimum and maximum F2 dimension number.}
\usage{
lowWAFOMSobol.dimF2MinMax(dimR)
}
\arguments{
\item{dimR}{dimension.}
}
\value{
supported minimum and maximum F2 dimension number
}
\description{
get minimum and maximum F2 dimension number.
}
