% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowWAFOMSobol.R
\name{lowWAFOMSobol.dimMinMax}
\alias{lowWAFOMSobol.dimMinMax}
\title{get minimum and maximum dimension number of Low WAFOM Niederreiter-Xing
Sequence}
\usage{
lowWAFOMSobol.dimMinMax()
}
\value{
supported minimum and maximum dimension number.
}
\description{
get minimum and maximum dimension number of Low WAFOM Niederreiter-Xing
Sequence
}
